Lots of changes (_exit, wait for all resources, code splitup)
Wait for all resources to be loaded before starting to find the CSS rules.

Use _exit() instead of exit() as the code cleanup WebKit seg fault seems to
happen in the atexit() handler.

Split the parsing/css lookup code in several functions. This will make it
easier to deal with @import.