Add deck2pdf
Saves a Deck JS presentation as a PDF.

This program is less polished than s5pdf but it works well enough in order
to build a PDF. Later on it should be merged with s5pdf.

Futhermore it would be good to disable or detect the end of the animations
and to implement 'steps' in case where a slide would have sub sections.